Hyperion.Config

Documentation

data HyperionConfig

Global configuration for Hyperion cluster.

Constructors

HyperionConfig

Fields

defaultSbatchOptions :: SbatchOptions

Default options to use for sbatch submissions

maxSlurmJobs :: Maybe Int

Maximum number of jobs to submit at a time

dataDir :: FilePath

Base directory for working dirs produced by newWorkDir

logDir :: FilePath

Base directory for all the log files

databaseDir :: FilePath

Base directory for databases

execDir :: FilePath

Base directory for copies of the main executable

jobDir :: FilePath

Base directory for SLURM job files

hyperionCommand :: Maybe FilePath

The command to run the main executable. Automatic if Nothing (see newClusterEnv)

initialDatabase :: Maybe FilePath

The database from which to initiate the program database

sshRunCommand :: SSHCommand

The command used to run ssh on nodes. Usually can be safely set to Nothing. See SSHCommand for details.

emailAddr :: Maybe Text

Email address for cluster notifications from hyperion. Nothing means no emails will be sent. Note that this setting can be different from the one in defaultSbatchOptions, which controls notifications from SLURM.

defaultHyperionConfig :: FilePath -> HyperionConfig

Default configuration, with all paths built form a single baseDirectory

newClusterEnv :: HyperionConfig -> HoldMap -> Int -> IO (ClusterEnv, FilePath)

Takes HyperionConfig and returns ClusterEnv, the path to the executable, and a new 'HoldMap.

Things to note:

programId is generated randomly.

If hyperionCommand is specified in HyperionConfig, then hyperionExec == hyperionCommand. Otherwise the running executable is copied to execDir with a unique name, and that is used as hyperionExec.

newDatabasePath is used to determine programDatabase from initialDatabase and databaseDir, programId.

timedProgramDir is used to determine programLogDir and programDataDir from the values in HyperionConfig and programId.

slurmWorkerLauncher is used for clusterWorkerLauncher

clusterDatabaseRetries is set to defaultDBRetries.

newDatabasePath :: Maybe FilePath -> FilePath -> ProgramId -> IO FilePath

Returns the path to a new database, given Maybe inital database filepath and base directory

If ProgramId id is XXXXX and initial database filename is original.sqlite, then the new filename is original-XXXXX.sqlite. If initial database path is Nothing, then the filename is XXXXX.sqlite.

The path is in subdirectory YYYY-mm (determined by current date) of base directory.

If inital database is given, then the new database is initilized with its contents.

timedProgramDir :: FilePath -> ProgramId -> IO FilePath

Given base directory and ProgramId (==XXXXX), returns the YYYY-mm/XXXXX subdirectory of the base directory (determined by current date).