Hyperion.Database.HasDB

General comments

Hyperion.Database.HasDB provides typeclass HasDB which describes environments that contain a DatabaseConfig, which is extracted by Lens' dbConfigLens.

This is used in the following way: if we have a monad m that

is an instance of MonadIO, i.e. embeds IO actions,

an instance of MonadReader, i.e. it carries an environment,

this environment is an instance of HasDB,

then we can create m-actions using withConnection, i.e.

doStuffWithConnection :: Sql.Connection -> IO a
...
do -- here we are in m monad
  ...
  result <- withConnection doStuffWithConnection
  ... 

withConnection uses Data.Pool. See withResource for details.

Documentation

data DatabaseConfig

Database information datatype

Constructors

DatabaseConfig

Fields

dbPool :: Pool Connection

dbProgramId :: ProgramId

dbRetries :: Int

class HasDB env where

HasDB typeclass

Minimal complete definition

dbConfigLens

Methods

dbConfigLens :: Lens' env DatabaseConfig

type Pool = Pool Connection

newDefaultPool :: FilePath -> IO (Pool Connection)

Produces a default pool with connections to the SQLite DB in the given file

withConnection :: forall m env a. (MonadIO m, MonadReader env m, HasDB env) => (Connection -> IO a) -> m a

Extracts the connection pool from the environment of our monad, gets a connection and runs the supplied function with it

withConnectionRetry :: forall m env a. (MonadIO m, MonadReader env m, HasDB env, MonadCatch m) => (Connection -> IO a) -> m a

Tries withConnection until succeeds. Failure means that SQLError is thrown during execution of the function. Otherwise execution is deemed successful. The number of attempts is determined by DatabaseConfig in the environment. If last attempt is a failure, the last exception propagates outside of withConnectionRetry. Uses retryRepeated internally.