Hyperion.ExtVar

Description

An ExtVar is an MVar that can be accessed by an external client. The "host" is the machine where the underlying MVar exists. The host can continue to use the underlying MVar as usual. A client can interact with it via functions like takeExtVar, putExtVar, readExtVar, etc., which behave in the same way as their MVar counterparts. An ExtVar can be recontstructed from its representation as a String or serialized to/from Binary data (and hence sent across a network).

For an example of using an ExtVar as a client, look in the hosts logs for a line that looks like:

Thu 010622 13:04:17

Made ExtVar: extVar @Int "login1.cm.cluster:39443:0" "test"

This shows that the host machine has made an ExtVar and it is ready to be accessed by a client. Now in a GHCi session (possibly on a completely different machine), you can do:

eVar = extVar @Int "login1.cm.cluster:39443:0" "test"
tryReadExtVarIO eVar

Just 42 > modifyExtVarIO_ eVar (x -> pure (x+1)) () > tryReadExtVarIO eVar Just 43

Documentation

data ExtVar a

extVar

Arguments

:: ByteString
End point address.

-> String
Name of the ExtVar

-> ExtVar a

ExtVar from an address and a name. To see what arguments you should pass to extVar, it is best to look for the "Made ExtVar" entry in the log o the host machine.

newExtVar :: (Binary a, Typeable a) => String -> MVar a -> Process (ExtVar a)

Make a new ExtVar from an MVar, together with a name. The host program can continue to use the MVar as usual. If the name is not unique, a ProcessRegistrationException will be thrown.

killExtVar :: forall a. (Binary a, Typeable a) => ExtVar a -> Process ()

Kill the server underlying the ExtVar. Subsequent calls from clients may block indefinitely.

takeExtVar :: (Binary a, Typeable a) => ExtVar a -> Process a

takeExtVar, etc. are analogous to takeMVar, etc. All functions block until they receive a response from the host.

tryTakeExtVar :: (Binary a, Typeable a) => ExtVar a -> Process (Maybe a)

putExtVar :: (Binary a, Typeable a) => ExtVar a -> a -> Process ()

tryPutExtVar :: (Binary a, Typeable a) => ExtVar a -> a -> Process Bool

readExtVar :: (Binary a, Typeable a) => ExtVar a -> Process a

tryReadExtVar :: (Binary a, Typeable a) => ExtVar a -> Process (Maybe a)

withExtVar :: (Binary a, Typeable a) => ExtVar a -> (a -> Process b) -> Process b

modifyExtVar_ :: (Binary a, Typeable a) => ExtVar a -> (a -> Process a) -> Process ()

modifyExtVar :: (Binary a, Typeable a) => ExtVar a -> (a -> Process (a, b)) -> Process b

takeExtVarIO :: (Binary a, Typeable a) => ExtVar a -> IO a

IO versions of ExtVar functions, for convenience.

tryTakeExtVarIO :: (Binary a, Typeable a) => ExtVar a -> IO (Maybe a)

putExtVarIO :: (Binary a, Typeable a) => ExtVar a -> a -> IO ()

tryPutExtVarIO :: (Binary a, Typeable a) => ExtVar a -> a -> IO Bool

readExtVarIO :: (Binary a, Typeable a) => ExtVar a -> IO a

tryReadExtVarIO :: (Binary a, Typeable a) => ExtVar a -> IO (Maybe a)

withExtVarIO :: (Binary a, Typeable a) => ExtVar a -> (a -> IO b) -> IO b

modifyExtVarIO_ :: (Binary a, Typeable a) => ExtVar a -> (a -> IO a) -> IO ()

modifyExtVarIO :: (Binary a, Typeable a) => ExtVar a -> (a -> IO (a, b)) -> IO b