Hyperion.HasWorkers

Documentation

class MonadBaseControl Process m => HasWorkers m where

Minimal complete definition

getWorkerLauncher

Methods

getWorkerLauncher :: m (WorkerLauncher JobId)

class HasWorkerLauncher env where

Minimal complete definition

toWorkerLauncher

Methods

toWorkerLauncher :: env -> WorkerLauncher JobId

withRemoteRun :: (Binary a, Typeable a, HasWorkers m) => ((SerializableClosureProcess (Either String a) -> Process a) -> m b) -> m b

remoteBind :: (Binary a, Typeable a, Binary b, Typeable b, HasWorkers m, StM m (SerializableClosureProcess (Either String b)) ~ SerializableClosureProcess (Either String b), StM m a ~ a) => m a -> StaticPtr (RemoteFunction a b) -> m b

remoteEval :: (Binary a, Typeable a, Binary b, Typeable b, HasWorkers m, StM m (SerializableClosureProcess (Either String b)) ~ SerializableClosureProcess (Either String b), StM m a ~ a) => StaticPtr (RemoteFunction a b) -> a -> m b

Orphan instances

MonadBase Process Process

Instance details

Methods

liftBase :: Process α -> Process α

MonadBaseControl Process Process

Instance details

Associated Types

type StM Process a :: *

Methods

liftBaseWith :: (RunInBase Process Process -> Process a) -> Process a

restoreM :: StM Process a -> Process a