Hyperion.HoldServer

Documentation

type HoldApi = ("retry" :> (Capture "service" Text :> Get '[JSON] (Maybe Text))) :<|> (("retry-all" :> Get '[JSON] [Text]) :<|> ("list" :> Get '[JSON] [Text]))

newtype HoldMap

Constructors

HoldMap (TVar (Map Text (MVar ())))

newHoldMap :: IO HoldMap

server :: HoldMap -> Server HoldApi

blockUntilRetried :: MonadIO m => HoldMap -> Text -> m ()

Start a hold associated to the given service. Returns an IO action that blocks until the hold is released

withHoldServer :: HoldMap -> (Int -> IO a) -> IO a

Start the hold server on an available port and pass the port number to the given action. The server is killed after the action finishes.