Hyperion.Remote

Types

newtype ServiceId

Type for service id. ServiceId is typically a random string that is assigned to a worker. (Maybe to other things too?)

Constructors

ServiceId String

serviceIdToText :: ServiceId -> Text

serviceIdToString :: ServiceId -> String

data WorkerMessage

Type for basic master to worker messaging

Constructors

Connected

ShutDown

data RemoteError

Constructors

RemoteError ServiceId RemoteErrorType

data RemoteErrorType

Detailed type for RemoteError. The constructors correspond to various possible AsyncResults.

Constructors

RemoteAsyncFailed DiedReason

RemoteAsyncLinkFailed DiedReason

RemoteAsyncCancelled

RemoteAsyncPending

RemoteException String

data WorkerConnectionTimeout

Constructors

WorkerConnectionTimeout ServiceId

data WorkerLauncher j

WorkerLauncher type parametrized by a type for job id.

Constructors

WorkerLauncher

Fields

withLaunchedWorker :: forall b. NodeId -> ServiceId -> (j -> Process b) -> Process b

A function that launches a worker for the given ServiceId on the master NodeId and supplies its job id to the given continuation

connectionTimeout :: Maybe NominalDiffTime

Timeout for the worker to connect. If the worker is launched into a Slurm queue, it may take a very long time to connect. In that case, it is recommended to set connectionTimeout = Nothing.

onRemoteError :: forall b. RemoteError -> Process b -> Process b

A handler for RemoteErrors. onRemoteError can do things like blocking before rerunning the remote process, or simply rethrowing the error.

Functions for running a worker

runProcessLocal :: Process a -> IO a

Run a Process locally using the default RemoteTable. Additionally allows a return value for the Process.

runProcessLocalWithRT :: RemoteTable -> Process a -> IO a

Run a Process locally using the specified RemoteTable. Additionally allows a return value for the Process.

runProcessLocalWithRT_ :: RemoteTable -> Process () -> IO ()

Spawns a new local Control.Distributed.Process.Node and runs the given Process on it. Waits for the process to finish.

Binds to the first available port by specifying port 0.

getExternalHostName :: IO String

Get a hostname for the current machine that does not correspond to a local network address (127.* or 10.*)

addressToNodeId :: Text -> NodeId

Convert a Text representation of EndPointAddress to NodeId. The format for the end point address is "TCP host:TCP port:endpoint id"

nodeIdToAddress :: NodeId -> Text

Inverse to addressToNodeId

masterNodeIdLabel :: String

masterNodeIdStatic :: Static (Maybe NodeId)

getMasterNodeId :: Process (Maybe NodeId)

registerMasterNodeId :: Maybe NodeId -> RemoteTable -> RemoteTable

initWorkerRemoteTable :: Maybe NodeId -> RemoteTable

worker

Arguments

:: NodeId
NodeId of the master node

-> ServiceId
ServiceId of master Process (should be registered)

-> Process ()

The main worker process.

Repeatedly (at most 5 times) send our own ProcessId and the send end of a typed channel (SendPort WorkerMessage) to a master node until it replies Connected (timeout 10 seconds for each attempt). Then expect a ShutDown signal.

While waiting, other processes will be run in a different thread, invoked by master through our NodeId (which it extracts from ProcessId)

withServiceId :: (ServiceId -> Process a) -> Process a

Registers (register) the current process under a random ServiceId, then passes the ServiceId to the given continuation. After the continuation returns, unregisters (unregister) the ServiceId.

withService :: Show j => WorkerLauncher j -> (NodeId -> ServiceId -> Process a) -> Process a

Start a new remote worker using WorkerLauncher and call a continuation with the NodeId and ServiceId of that worker. The continuation is run in the process that is registered under the ServiceId (see withServiceId).

Throws (throwM) a WorkerConnectionTimeout if worker times out (timeout described in WorkerLauncher)

The call to the user function is bracketed by worker startup and shutdown procedures.

Functions related to running remote functions

data SerializableClosureProcess a

A process that generates the Closure to be run on a remote machine.

Constructors

SerializableClosureProcess

Fields

runClosureProcess :: Process (Closure (Process (Either String a)))

Process to generate a closure. This process will be run when a remote location has been identified that the closure can be sent to.

staticSDict :: Closure (SerializableDict (Either String a))

Dict for seralizing the result.

closureVar :: MVar (Closure (Process (Either String a)))

If a remote computation fails, it may be added to the HoldMap to be tried again. In that case, we don't want to evaluate runClosureProcess again, so we use an MVar to memoize the result of runClosureProcess.

getClosure :: SerializableClosureProcess a -> Process (Closure (Process (Either String a)))

Get closure and memoize the result

type RemoteProcessRunner = forall a. (Binary a, Typeable a) => SerializableClosureProcess a -> Process a

The type of a function that takes a SerializableClosureProcess and runs the Closure on a remote machine. In case the remote machine returns a Left value (i.e. an error), throws this value wrapped in RemoteError of type RemoteException. May throw other RemoteErrors if remote execution fails in any way.

withRemoteRunProcess :: Show j => WorkerLauncher j -> (RemoteProcessRunner -> Process a) -> Process a

Starts a new remote worker and runs a user function, which is supplied with RemoteProcessRunner for running closures on that worker. If a RemoteError occurs, it is handled by the onRemoteError supplied in the WorkerLauncher.

tryLogException :: Process b -> Process (Either String b)

Catch any exception, log it, and return as a string. In this way, errors will be logged by the worker where they occurred, and also sent up the tree.

mkSerializableClosureProcess :: Typeable b => Closure (Dict (Serializable b)) -> Process (Closure (Process b)) -> Process (SerializableClosureProcess b)

Construct a SerializableClosureProcess for evaluating the closure mb on a remote machine. The action mb that produces the Closure will only be run once -- when a worker first becomes available. The MVar v caches the resulting Closure (so it can be re-used in the event of an error and retry), which will then be sent across the network.

Note that we wrap the given closure in tryLogException. Thus, exception handling is added by catching any exception e, logging e with err (on the worker), and returning a Left result with the textual representation of e from Show instance.