Hyperion.Util

Documentation

randomString :: Int -> IO String

IO action that returns a random string of given length

retryRepeated

Arguments

:: (Show e, MonadIO m)

=> Int
If this is 0 (or less), then it attempt doTry m indefinitely.

-> (m a -> m (Either e a))

-> m a

-> m a

retryRepeated n doTry m tries to run doTry m n-1 times, after which it runs m 1 time. After each failure waits 15-90 seconds randomly. Returns on first success. Failure is represented by a Left value.

shellEsc :: FilePath -> [String] -> String

Takes a path and a list of String arguments, shell-escapes the arguments, and combines everything into a single string.

onceWithArgument :: (MonadIO m, MonadIO n) => (m a -> n b) -> m a -> n (a, b)

Given a function that takes a monadic action, allow us to inspect the argument to that function. Only works if f is guaranteed to evaluate its argument exactly once. This function is useful in combination with remoteBind, since it allows one to inspect the argument produced by the monadic action passed to the given continuation.

minute :: NominalDiffTime

hour :: NominalDiffTime

day :: NominalDiffTime

nominalDiffTimeToMicroseconds :: NominalDiffTime -> Int

myExecutable :: IO FilePath

savedExecutable

Arguments

:: FilePath

-> String
the string to append

-> IO FilePath

Determine the path to this executable and save a copy with a string appended

sanitizeFileString :: String -> FilePath

Replaces all non-allowed characters by '_'. Allowed characters are alphanumerics and .,-,_

hashTruncateString :: Int -> String -> String

Truncates a string to a string of at most given length, replacing dropped characters by a hash. The hash takes up 11 symbols (depends on 'maxBound'::'Int'), so asking for a smaller length will not work.

hashTruncateFileName :: String -> String

Synonim for hashTruncateString 230

Tuples

These currying/uncurrying routines are useful because remote functions can only take a single argument

curry3 :: ((a, b, c) -> d) -> a -> b -> c -> d

Converts an uncurried function to a curried function.

uncurry3 :: (a -> b -> c -> d) -> (a, b, c) -> d

Converts a curried function to a function on a triple.

curry4 :: ((a, b, c, d) -> e) -> a -> b -> c -> d -> e

Converts an uncurried function to a curried function.

uncurry4 :: (a -> b -> c -> d -> e) -> (a, b, c, d) -> e

Converts a curried function to a function on a quadruple.