Hyperion.HasWorkers

Documentation

class MonadBaseControl Process m => HasWorkers m where

A class for monads that can run things in the Process monad, and have access to a WorkerLauncher. An instance of HasWorkers can use remoteBind and remoteEval to run computations in worker processes at remote locations.

Minimal complete definition

getWorkerLauncher

Methods

getWorkerLauncher :: m (WorkerLauncher JobId)

class HasWorkerLauncher env where

A class indicating that type env contains a WorkerLauncher.

Minimal complete definition

toWorkerLauncher

Methods

toWorkerLauncher :: env -> WorkerLauncher JobId

withRemoteRun :: HasWorkers m => (RemoteProcessRunner -> m a) -> m a

Uses the WorkerLauncher to get a RemoteProcessRunner and pass it to the given continuation.

This function is essentially a composition of getWorkerLauncher with withRemoteRunProcess, lifted from Process to m using MonadBaseControl.

We use the machinery of MonadBaseControl because withRemoteRunProcess expects something that runs in the Process monad, not in m. Our main use case is when 'm ~ ReaderT env Process', where env is an instance of HasWorkerLauncher. In that case, we would like to capture the env at the beginning, and then use flip runReaderT env to get from m to Process and lift to get from Process back to m. This is what the MonadBaseControl instance for ReaderT does.

remoteBind :: (Binary a, Typeable a, Binary b, Typeable b, HasWorkers m, StM m (SerializableClosureProcess (Either String b)) ~ SerializableClosureProcess (Either String b), StM m a ~ a) => m a -> StaticPtr (RemoteFunction a b) -> m b

Evaluate ma to get an argument, and pass that argument to the given RemoteFunction, evaluating the result at a remote location. The result is automatically serialized and sent back over the wire. If you squint and replace StaticPtr (RemoteFunction a b) with a -> m b, then the type becomes m a -> (a -> m b) -> m b, which is the type of bind (>>=); hence the name remoteBind.

This function essentially uses withRemoteRun to get a RemoteProcessRunner, and bindRemoteStatic to produce a SerializableClosureProcess that is fed to the RemoteProcessRunner. All lifted to m.

remoteEval :: (Binary a, Typeable a, Binary b, Typeable b, HasWorkers m, StM m (SerializableClosureProcess (Either String b)) ~ SerializableClosureProcess (Either String b), StM m a ~ a) => StaticPtr (RemoteFunction a b) -> a -> m b

Evaluate a RemoteFunction on the given argument at a remote location. The result is automatically serialized and sent back over the wire. If you squint and replace 'StaticPtr (RemoteFunction a b)' with 'a -> m b', then the type becomes '(a -> m b) -> a -> m b', which is just function application.

Shorthand for remoteBind appopriately composed with return :: a -> m a.

Orphan instances

MonadBase Process Process
Trivial orphan instance of MonadBase for Process.

Instance details

Methods

liftBase :: Process α -> Process α

MonadBaseControl Process Process
Trivial orphan instance of MonadBaseControl for Process

Instance details

Associated Types

type StM Process a :: *

Methods

liftBaseWith :: (RunInBase Process Process -> Process a) -> Process a

restoreM :: StM Process a -> Process a