Hyperion.HasWorkers

Documentation

class (MonadBase Process m, MonadUnliftProcess m, MonadIO m) => HasWorkers m where

A class for monads that can run things in the Process monad, and have access to a WorkerLauncher. An instance of HasWorkers can use remoteBind and remoteEval to run computations in worker processes at remote locations.

Methods

getWorkerLauncher :: m (WorkerLauncher JobId)

class MonadUnliftProcess m where

A class for Monads that can run continuations in the Process monad, modeled after MonadUnliftIO (https:/hackage.haskell.orgpackageunliftio-core-0.2.0.1docs/Control-Monad-IO-Unlift.html).

Methods

withRunInProcess :: ((forall a. m a -> Process a) -> Process b) -> m b

class HasWorkerLauncher env where

A class indicating that type env contains a WorkerLauncher.

Methods

toWorkerLauncher :: env -> WorkerLauncher JobId

withRemoteRun :: HasWorkers m => (RemoteProcessRunner -> m a) -> m a

Uses the WorkerLauncher to get a RemoteProcessRunner and pass it to the given continuation.

This function is essentially a composition of getWorkerLauncher with withRemoteRunProcess, lifted from Process to m using MonadUnliftProcess.

We use the machinery of MonadUnliftProcess because withRemoteRunProcess expects something that runs in the Process monad, not in m. Our main use case is when 'm ~ ReaderT env Process', where env is an instance of HasWorkerLauncher.

remoteEvalWithDictM :: (HasWorkers m, Serializable b) => Closure (Dict (Serializable b)) -> m (Closure (Process b)) -> m b

Compute a closure at a remote location. The user supplies an 'm (Closure (...))' which is only evaluated when a remote worker becomes available (for example after the worker makes it out of the Slurm queue).

remoteEvalM :: (HasWorkers m, Static (Binary b), Typeable b) => m (Closure (Process b)) -> m b

Evaluate a Closure at a remote location, assuming a 'Static (Binary b)' instance. The Closure itself is ony computed when a worker becomes available.

remoteEval :: (HasWorkers m, Static (Binary b), Typeable b) => Closure (Process b) -> m b

Evaluate a Closure at a remote location.

Orphan instances

MonadBase Process Process
Trivial orphan instance of MonadBase for Process.

Instance details

Methods

liftBase :: Process α -> Process α