Hyperion.Remote

Types

newtype ServiceId

Type for service id. ServiceId is typically a random string that is assigned to a worker. (Maybe to other things too?)

Constructors

ServiceId String

serviceIdToText :: ServiceId -> Text

serviceIdToString :: ServiceId -> String

data WorkerMessage

Type for basic master to worker messaging

Constructors

Connected

ShutDown

data RemoteError

Constructors

RemoteError ServiceId RemoteErrorType

data RemoteErrorType

Detailed type for RemoteError. The constructors correspond to various possible AsyncResults.

Constructors

RemoteAsyncFailed DiedReason

RemoteAsyncLinkFailed DiedReason

RemoteAsyncCancelled

RemoteAsyncPending

RemoteException String

data WorkerConnectionTimeout

Constructors

WorkerConnectionTimeout ServiceId

data WorkerLauncher j

WorkerLauncher type parametrized by a type for job id.

Constructors

WorkerLauncher

Fields

withLaunchedWorker :: forall b. NodeId -> ServiceId -> (j -> Process b) -> Process b

A function that launches a worker for the given ServiceId on the master NodeId and supplies its job id to the given continuation

connectionTimeout :: Maybe NominalDiffTime

Timeout for the worker to connect. If the worker is launched into a Slurm queue, it may take a very long time to connect. In that case, it is recommended to set connectionTimeout = Nothing.

serviceHoldMap :: Maybe HoldMap

The HoldMap which we can use to block on errors. If set to Nothing, then the Hyperion.HoldServer functionality won't be used.

Functions for running a worker

runProcessLocallyDefault :: Process a -> IO a

runProcessLocally with default RemoteTable and trying ports [10090 .. 10990]. See runProcessLocally and runProcessLocally_.

runProcessLocally :: RemoteTable -> [ServiceName] -> Process a -> IO a

Same as runProcessLocally_, but returns allows a return value for the Process.

runProcessLocally_ :: RemoteTable -> [ServiceName] -> Process () -> IO ()

Spawns a new local Control.Distributed.Process.Node with the given RemoteTable and runs the given Process on it. Waits for the process to finish.

It tries to bind the node to the given list [ServiceName] of ports, attempting them one-by-one, and waiting for 5 seconds before timing the port out.

addressToNodeId :: Text -> NodeId

Convert a Text representation of EndPointAddress to NodeId. The format for the end point address is "TCP host:TCP port:endpoint id"

nodeIdToAddress :: NodeId -> Text

Inverse to addressToNodeId

worker

Arguments

:: NodeId
NodeId of the master node

-> ServiceId
ServiceId of master Process (should be registered)

-> Process ()

The main worker process.

Repeatedly (at most 5 times) send our own ProcessId and the send end of a typed channel (SendPort WorkerMessage) to a master node until it replies Connected (timeout 10 seconds for each attempt). Then expect a ShutDown signal.

While waiting, other processes will be run in a different thread, invoked by master through our NodeId (which it extracts from ProcessId)

withServiceId :: (ServiceId -> Process a) -> Process a

Registers (register) the current process under a random ServiceId, then runs the given function on this ServiceId. After the function returns, unregisters (unregister) the ServiceId.

withService :: Show j => WorkerLauncher j -> (NodeId -> ServiceId -> Process a) -> Process a

Start a new remote worker using WorkerLauncher and call a function with the NodeId and ServiceId of that worker.

Throws (throwM) a WorkerConnectionTimeout if worker times out (timeout described in WorkerLauncher)

The call to the user function is bracketed by worker startup and shutdown procedures.

Functions related to running remote functions

data SerializableClosureProcess a

A process that generates the Closure to be run on a remote machine.

Constructors

SerializableClosureProcess

Fields

runClosureProcess :: Process (Closure (Process a))

Process to generate closure

staticSDict :: Static (SerializableDict a)

Dict for serializing result

closureVar :: MVar (Closure (Process a))

MVar for memoizing closure so we don't run process more than once

getClosure :: SerializableClosureProcess a -> Process (Closure (Process a))

Get closure and memoize the result

type RemoteProcessRunner = forall a. (Binary a, Typeable a) => SerializableClosureProcess (Either String a) -> Process a

The type of a function that takes a SerializableClosureProcess and runs the Closure on a remote machine. In case the remote machine returns a Left value (i.e. an error), throws this value wrapped in RemoteError of type RemoteException. May throw other RemoteErrors if remote execution fails in any way.

withRemoteRunProcess :: Show j => WorkerLauncher j -> (RemoteProcessRunner -> Process a) -> Process a

Starts a new remote worker and runs a user function, which is supplied with RemoteProcessRunner for running closures on that worker. If WorkerLauncher has a HoldMap, then RemoteError exceptions propagating out of the supplied user function are logged and withRemoteRunProcess blocks until the ServiceId is released by HoldServer. Then withRemoteRunProcess restarts (generating new ServiceId). If no HoldMap is available, then the exception propagates out of withRemoteRunProcess.

data RemoteFunction a b

A monadic function, together with SerializableDicts for its input and output. In the output type 'Either String b', String represents an error in remote execution.

Constructors

RemoteFunction

Fields

remoteFunctionRun :: a -> Process (Either String b)

sDictIn :: SerializableDict a

sDictOut :: SerializableDict (Either String b)

remoteFn :: (Binary a, Typeable a, Binary b, Typeable b) => (a -> Process b) -> RemoteFunction a b

Produces a RemoteFunction from a monadic function. Exception handling is added by catching any exception e, logging e with err (on the worker), and returning a Left result with the textual representation of e from Show instance.

remoteFnIO :: (Binary a, Typeable a, Binary b, Typeable b) => (a -> IO b) -> RemoteFunction a b

Same as remoteFn but takes a function in IO monad.

remoteFunctionRunStatic :: (Typeable a, Typeable b) => Static (RemoteFunction a b -> a -> Process (Either String b))

remoteFunctionRun of RemoteFunction wrapped in Static.

sDictInStatic :: (Typeable a, Typeable b) => Static (RemoteFunction a b -> SerializableDict a)

sDictIn of RemoteFunction wrapped in Static.

sDictOutStatic :: (Typeable a, Typeable b) => Static (RemoteFunction a b -> SerializableDict (Either String b))

sDictOut of RemoteFunction wrapped in Static.

bindRemoteStatic :: (Binary a, Typeable a, Typeable b) => Process a -> StaticPtr (RemoteFunction a b) -> Process (SerializableClosureProcess (Either String b))

Constructs SerializableClosureProcess given an action that constructs the argument to the RemoteFunction, and a StaticPtr to the remote function. The action that constructs the argument is embedded into runClosureProcess of the resulting SerializableClosureProcess.

StaticPtr to the remote function can be constructed by using the keyword 'static' from StaticPointers extention. E.g.,

static $ remoteFn f 

Note, however, that f should be either a top-level definition or a closed local definition in order for 'static' to work. Closed local definitions are local definitions that could in principle be defined at top-level. A necessary condition is that the whole object to which 'static' is applied to is known at compile time.

applyRemoteStatic :: (Binary a, Typeable a, Typeable b) => StaticPtr (RemoteFunction a b) -> a -> Process (SerializableClosureProcess (Either String b))

Same as bindRemoteStatic where the argument to RemoteFunction is constructed by return from the supplied argument.