Hyperion.Slurm.Sbatch

Documentation

data SbatchError

Error from running sbatch. The Strings are the contents of stdout and stderr from sbatch.

Constructors

SbatchError (ExitCode, String, String)

data SbatchOptions

Type representing possible options for sbatch. Map 1-to-1 to sbatch options, so see man sbatch for details.

Constructors

SbatchOptions

Fields

jobName :: Maybe Text

Job name ("--job-name")

chdir :: Maybe FilePath

Working directory for the job ("--D")

output :: Maybe FilePath

Where to direct stdout of the job ("--output")

nodes :: Int

Number of nodes ("--nodes")

nTasksPerNode :: Int

Number of tasks per node ("--ntasks-per-node")

time :: NominalDiffTime

Job time limit ("--time")

mem :: Maybe Text

Memory per node, use suffix K,M,G, or T to define the units. ("--mem")

mailType :: Maybe Text

("--mail-type")

mailUser :: Maybe Text

("--mail-user")

partition :: Maybe Text

SLURM partition ("--partition")

defaultSbatchOptions :: SbatchOptions

Default SbatchOptions. Request 1 task on 1 node for 24 hrs, everything else unspecified.

sBatchOptionString :: SbatchOptions -> String

sbatchOutputParser :: Parser JobId

sbatchScript :: SbatchOptions -> String -> IO JobId

Runs sbatch on a batch file with options pulled from SbatchOptions and script given as the String input parameter. If sbatch exists with failure then throws SbatchError.

formatRuntime :: NominalDiffTime -> String

Formats NominalDiffTime into hh:mm:ss.

sbatchCommand :: SbatchOptions -> FilePath -> [Text] -> IO JobId

Runs the command given by FilePath with arguments [Text] in sbatch script via sbatchScript. If sbatchScript throws SbatchError, retries for a total of 3 attempts (via retryRepeated).