Hyperion.Util

Documentation

newtype Unique

An opaque type representing a unique object. Only guaranteed to be unique in one instance of a running program. For example, if we allowed Unique's to be serialized and sent across the wire, or stored and retrieved from a database, they would no longer be guaranteed to be unique.

Constructors

MkUnique Integer

uniqueSource :: IORef Integer

newUnique :: IO Unique

Get a new Unique.

randomString :: Int -> IO String

IO action that returns a random string of given length

retryRepeated

Arguments

:: (Show e, MonadIO m)

=> Int
If this is 0 (or less), then it attempt doTry m indefinitely.

-> (m a -> m (Either e a))

-> m a

-> m a

retryRepeated n doTry m tries to run doTry m n-1 times, after which it runs m 1 time. After each failure waits 15-90 seconds randomly. Returns on first success. Failure is represented by a Left value.

data WaitRetry e

Constructors

WaitRetry

Fields

err :: e

waitTime :: Int

retryExponential :: MonadIO m => (m a -> m (Either e a)) -> (WaitRetry e -> m ()) -> m a -> m a

retryExponential doTry m tries to run doTry m. After the n-th successive failure, it waits time 2^n*t0, where t0 is a randomly chosen time between 10 and 20 seconds. Unlike retryRepeated, retryExponential never eventually throws an exception, so it should only be used when the only way to recover from the exception without the whole program crashing is to retry until things work. Typically this means it should only be used in the master process.

emailMessage :: MonadIO m => Text -> Text -> Text -> m ()

Send an email to toAddr with the given subject and body.

email :: (Show a, MonadIO m) => Text -> Text -> a -> m ()

Send an email to toAddr showing the object a. The subject line is "msg: ...", where msg is the first argument and "..." is the first 40 characters of 'show a'.

emailError :: (Show a, MonadIO m) => Text -> a -> m ()

Send an email with msg Error

shellEsc :: FilePath -> [String] -> String

Takes a path and a list of String arguments, shell-escapes the arguments, and combines everything into a single string.

minute :: NominalDiffTime

hour :: NominalDiffTime

day :: NominalDiffTime

nominalDiffTimeToMicroseconds :: NominalDiffTime -> Int

myExecutable :: IO FilePath

savedExecutable

Arguments

:: FilePath

-> String
the string to append

-> IO FilePath

Determine the path to this executable and save a copy to the specified dir with a string appended to filename.

sanitizeFileString :: String -> FilePath

Replaces all non-allowed characters by '_'. Allowed characters are alphanumerics and .,-,_

hashTruncateString :: Int -> String -> String

Truncates a string to a string of at most given length, replacing dropped characters by a hash. The hash takes up 43 symbols, so asking for a smaller length will still return 43 symbols.

hashTruncateFileName :: String -> String

Synonim for hashTruncateString 230

withDict :: forall (c :: Constraint) r. (c => r) -> Dict c -> r

Turn an expression with a constraint into a function of an explicit dictionary